--- src/main/java/dev/task/dndquest/model/entity/advententity/AdventEntity.java
@@ -1,10 +1,0 @@
-package dev.task.dndquest.model.entity.advententity;
-
-import jakarta.persistence.*;
-
-@MappedSuperclass
-public abstract class AdventEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-}
