+++ src/main/java/dev/task/dndquest/model/entity/adventure/Interaction.java
@@ -1,4 +1,4 @@
-package dev.task.dndquest.model.entity;
+package dev.task.dndquest.model.entity.adventure;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
