+++ src/main/java/dev/task/dndquest/model/entity/character/PlayCharacterClass.java
@@ -1,4 +1,4 @@
-package dev.task.dndquest.model.entity;
+package dev.task.dndquest.model.entity.character;
 
 import jakarta.persistence.*;
 import lombok.Data;
@@ -7,8 +7,8 @@ import lombok.NoArgsConstructor;
 @Entity
 @Data
 @NoArgsConstructor
-@Table(name = "race")
-public class Race {
+@Table(name = "class")
+public class PlayCharacterClass {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
