+++ src/main/java/dev/task/dndquest/model/entity/character/Race.java
@@ -1,0 +1,16 @@
+package dev.task.dndquest.model.entity.character;
+
+import jakarta.persistence.*;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Data
+@NoArgsConstructor
+@Table(name = "race")
+public class Race {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String name;
+}
