+++ src/main/java/dev/task/dndquest/model/entity/item/AdventEntity.java
@@ -1,0 +1,7 @@
+package dev.task.dndquest.model.entity.item;
+
+import jakarta.persistence.*;
+
+@MappedSuperclass
+public abstract class AdventEntity {
+}
