+++ src/main/java/dev/task/dndquest/repository/ItemRepository.java
@@ -1,6 +1,6 @@
 package dev.task.dndquest.repository;
 
-import dev.task.dndquest.model.entity.advententity.Item;
+import dev.task.dndquest.model.entity.item.Item;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
