+++ src/main/java/dev/task/dndquest/service/ItemService.java
@@ -1,6 +1,6 @@
 package dev.task.dndquest.service;
 
-import dev.task.dndquest.model.entity.advententity.Item;
+import dev.task.dndquest.model.entity.item.Item;
 
 public interface ItemService {
     Item findByName(String name);
