+++ src/main/java/dev/task/dndquest/service/PlayCharacterClassService.java
@@ -1,6 +1,6 @@
 package dev.task.dndquest.service;
 
-import dev.task.dndquest.model.entity.PlayCharacterClass;
+import dev.task.dndquest.model.entity.character.PlayCharacterClass;
 
 public interface PlayCharacterClassService {
     PlayCharacterClass findByName(String name);
