+++ src/main/java/dev/task/dndquest/service/PlayCharacterService.java
@@ -3,7 +3,7 @@ package dev.task.dndquest.service;
 import dev.task.dndquest.model.dto.request.InventoryRequestDto;
 import dev.task.dndquest.model.dto.request.PlayCharacterRequestDto;
 import dev.task.dndquest.model.dto.response.InventoryResponseDto;
-import dev.task.dndquest.model.entity.PlayCharacter;
+import dev.task.dndquest.model.entity.character.PlayCharacter;
 import java.util.List;
 
 public interface PlayCharacterService {
