+++ src/main/java/dev/task/dndquest/service/RaceService.java
@@ -1,6 +1,6 @@
 package dev.task.dndquest.service;
 
-import dev.task.dndquest.model.entity.Race;
+import dev.task.dndquest.model.entity.character.Race;
 
 public interface RaceService {
     Race findByName(String name);
