+++ src/main/java/dev/task/dndquest/service/impl/PlayCharacterServiceImpl.java
@@ -7,8 +7,8 @@ import dev.task.dndquest.model.dto.request.InventoryRequestDto;
 import dev.task.dndquest.model.dto.request.PlayCharacterRequestDto;
 import dev.task.dndquest.model.dto.response.InventoryResponseDto;
 import dev.task.dndquest.model.dto.response.PlayCharacterResponseDto;
-import dev.task.dndquest.model.entity.advententity.Item;
-import dev.task.dndquest.model.entity.PlayCharacter;
+import dev.task.dndquest.model.entity.item.Item;
+import dev.task.dndquest.model.entity.character.PlayCharacter;
 import dev.task.dndquest.repository.PlayCharacterRepository;
 import dev.task.dndquest.service.ItemService;
 import dev.task.dndquest.service.PlayCharacterClassService;
