+++ src/test/java/dev/task/dndquest/service/impl/PlayCharacterServiceTest.java
@@ -3,9 +3,9 @@ package dev.task.dndquest.service.impl;
 import dev.task.dndquest.mapper.DtoMapper;
 import dev.task.dndquest.model.dto.request.PlayCharacterRequestDto;
 import dev.task.dndquest.model.dto.response.PlayCharacterResponseDto;
-import dev.task.dndquest.model.entity.PlayCharacter;
-import dev.task.dndquest.model.entity.PlayCharacterClass;
-import dev.task.dndquest.model.entity.Race;
+import dev.task.dndquest.model.entity.character.PlayCharacter;
+import dev.task.dndquest.model.entity.character.PlayCharacterClass;
+import dev.task.dndquest.model.entity.character.Race;
 import dev.task.dndquest.repository.PlayCharacterRepository;
 import dev.task.dndquest.service.PlayCharacterClassService;
 import dev.task.dndquest.service.RaceService;
