+++ src/test/java/dev/task/dndquest/service/impl/RaceServiceTest.java
@@ -1,7 +1,7 @@
 package dev.task.dndquest.service.impl;
 
 import dev.task.dndquest.exception.RaceNotFoundException;
-import dev.task.dndquest.model.entity.Race;
+import dev.task.dndquest.model.entity.character.Race;
 import dev.task.dndquest.repository.RaceRepository;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeAll;
